move `FieldLink` component to `@graphiql/react`
